pyproject.toml
[project]
name = "cover-letter-writer-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "agno>=1.4.6",
    "groq>=0.24.0",
    "python-dotenv>=1.1.0",
]
